Separate declarations from implementation
